galileo-adk/pyproject.toml
[project]
name = "galileo-adk"
version = "0.3.0"
description = "Galileo observability integration for Google ADK"
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = { text = "Apache-2.0" }
authors = [{ name = "Galileo Technologies Inc.", email = "team@galileo.ai" }]
dependencies = [
    "galileo>=1.39.0,<2.0.0",
    "google-adk>=1.14.1,<2.0.0",
]

[project.urls]
Repository = "https://github.com/rungalileo/galileo-python"

# UV-specific configuration
[tool.uv]
sources = { galileo = { workspace = true } }

# Development dependencies (UV standard - use dependency-groups)
[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.7.0",
    "pytest-socket>=0.7",
    "vcrpy>=7.0.0",
    "ruff>=0.12.3",
    "mypy>=1.16.0",
    "coverage>=7.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/galileo_adk"]

# Test configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
env = [
    "GALILEO_CONSOLE_URL=http://localtest:8088",
    "GALILEO_API_KEY=api-1234567890",
    "GALILEO_PROJECT=test-project",
    "GALILEO_LOG_STREAM=test-log-stream",
]
addopts = [
    "-v",
    "--strict-markers",
    "--disable-warnings",
    "--showlocals",
]

# Linting and formatting
[tool.ruff]
line-length = 120
fix = true
src = ["src"]
target-version = "py310"

[tool.ruff.lint]
select = ["F", "E4", "E7", "E9", "W6", "I", "UP", "ASYNC", "B", "C4", "ARG", "PT", "RET", "SIM"]
ignore = [
    "S101",     # Use of assert (needed for tests)
    "ARG001",   # Unused function argument
    "ARG002",   # Unused method argument
    "D",        # Docstring rules (relaxed for MVP)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ANN",      # Type annotations (not critical for tests)
    "PLR",      # Complexity rules
    "S",        # Security rules (tests need assertions)
    "ARG",      # Unused arguments
    "D",        # Docstring rules
]

[tool.ruff.lint.isort]
known-first-party = ["galileo_adk", "galileo"]

# Type checking
[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
follow_imports = "skip"

# Coverage
[tool.coverage.run]
source = ["src/galileo_adk"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
